+++ src/main/resources/static/docs/index.html
@@ -455,6 +455,7 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <li><a href="#_로그인_api">로그인 API</a>
 <ul class="sectlevel2">
 <li><a href="#_성공적인_로그인">성공적인 로그인</a></li>
+<li><a href="#_실패_케이스_2">실패 케이스</a></li>
 </ul>
 </li>
 </ul>
@@ -800,6 +801,9 @@ Content-Length: 80
 }</code></pre>
 </div>
 </div>
+</div>
+<div class="sect2">
+<h3 id="_실패_케이스_2"><a class="link" href="#_실패_케이스_2">실패 케이스</a></h3>
 <div class="sect3">
 <h4 id="_유저가_없을_때"><a class="link" href="#_유저가_없을_때">유저가 없을 때</a></h4>
 <div class="listingblock">
@@ -865,7 +869,7 @@ Content-Length: 139
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2025-04-03 18:45:56 +0900
+Last updated 2025-04-03 18:48:41 +0900
 </div>
 </div>
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.15.6/styles/github.min.css">
