+++ src/test/java/com/barogo/java/delivery/poc/controller/AuthControllerTest.java
@@ -27,8 +27,8 @@ import com.barogo.java.delivery.poc.common.exception.BaseException;
 import com.barogo.java.delivery.poc.common.exception.GlobalExceptionHandler;
 import com.barogo.java.delivery.poc.dto.request.UserSigninRequest;
 import com.barogo.java.delivery.poc.dto.request.UserSignupRequest;
-import com.barogo.java.delivery.poc.service.UserCreator;
-import com.barogo.java.delivery.poc.service.UserFacade;
+import com.barogo.java.delivery.poc.service.user.UserCreator;
+import com.barogo.java.delivery.poc.service.user.UserFacade;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @ExtendWith({RestDocumentationExtension.class, MockitoExtension.class})
